requirements.txt
streamlit
PathLib
IPython
PyPDF2
google.generativeai
python-dotenv
